implemented gifenc
