+++ package.json
@@ -41,6 +41,7 @@
     "dotenv": "^10.0.0",
     "file-type": "^9.0.0",
     "get-image-colors": "^4.0.0",
+    "gifenc": "^1.0.3",
     "gifencoder": "^2.0.1",
     "gifwrap": "^0.9.2",
     "jimp": "^0.16.1",
